Avoid circular imports
circle: graph_endpoints -> graph-project-runner -> session.schema -> aura_gds -> graph_endpoints

By seperating the remote classes from the non-remotes, we break the circular import
